Move web.TestFetSelf > interceptor.TestUserVerifier
